Переделал программу, убрал старые классы...Пока не плюсует скидки и дисконты

+++ software_evolution/Item.cs
@@ -30,7 +30,7 @@ namespace software_evolution
             return _Goods;
         }
 
-        public int GetBonus()
+        public double GetBonus()
         {
             return _Goods.GetBonus(this);
         }
